Update 3:
That didn't work. Fixed it now though.
-----------------------
Fixed the check thingy.
------------------------------------------------------------------------------------------------------------
That's all for now. The next newsletter should be coming this year on Halloween, though this is speculative.

Here's why:

In the most recent newsletter (9th Anniversary; Sep. 2024; Autumn 2024), there is a line at the end that reads, "See you this year, and next year." This suggests that there will be another newsletter or event later this year, as well as something in 2025.

Given this, a Halloween 2024 newsletter seems likely, especially since it aligns with last year's Halloween 2023 newsletter. For next year, the most probable newsletters might be Winter 2025, pairing with the Winter 2023 newsletter, or possibly Spring 2025.